Move fetch calls to a service file
